requirements.txt
watchdog
requests
PyYAML
